move player-model related functions from bpl_interface into prediction_utils
